fix(frontend): account modal in mobile
